Remove unnecessary unboxing.

+++ Squeezer/src/main/java/uk/org/ngo/squeezer/dialog/ServerAddressPreference.java
@@ -16,8 +16,6 @@
 
 package uk.org.ngo.squeezer.dialog;
 
-import org.acra.ErrorReporter;
-
 import android.annotation.TargetApi;
 import android.app.ProgressDialog;
 import android.content.Context;
@@ -40,6 +38,8 @@ import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Spinner;
 import android.widget.TextView;
+
+import org.acra.ErrorReporter;
 
 import java.io.IOException;
 import java.net.DatagramPacket;
@@ -458,7 +458,7 @@ public class ServerAddressPreference extends DialogPreference {
          */
         @Override
         protected void onProgressUpdate(Integer... values) {
-            mPref.updateProgress(Math.min(mServerMap.size(), 5), values[0].intValue());
+            mPref.updateProgress(Math.min(mServerMap.size(), 5), values[0]);
         }
 
         @Override
